# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Apricode documentation",
        "description": "V1 Api docs",
        "contact": {
            "email": "ialeroy@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/game": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Список игр",
                "description": "Список игр",
                "operationId": "23a5c91cd094905106b2687c39e4ae5d",
                "parameters": [
                    {
                        "name": "page",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "count",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Game"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/Pagination"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Games"
                ],
                "summary": "Добавить игру",
                "description": "Добавить игру",
                "operationId": "2ae4807231f40493fe16d22d82a1eabb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Game"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/game/{id}": {
            "get": {
                "tags": [
                    "Games"
                ],
                "summary": "Страница игры",
                "description": "Страница игры",
                "operationId": "150eb0bc8ce8f88c17f5721c799eb367",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Game"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Games"
                ],
                "summary": "Удаление игры",
                "description": "Удаление игры",
                "operationId": "c81468c67014daffd5c29f32d5e2c6e2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "patch": {
                "tags": [
                    "Games"
                ],
                "summary": "Обновление игры",
                "description": "Обновление игры",
                "operationId": "d10718ce7225c57b7ae2bb216d90d6a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GameRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Game"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "422": {
                        "description": "Unprocessable entity.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Game": {
                "title": "Game",
                "description": "Game Model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "Имя",
                        "type": "string",
                        "example": "Rachelle Homenick"
                    },
                    "studio": {
                        "title": "studio",
                        "description": "Студия",
                        "type": "string",
                        "example": "Logic"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Создана в",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Game"
                }
            },
...
            "Pagination": {
                "title": "Pagination",
                "description": "Pagination",
                "properties": {
                    "current_page": {
                        "title": "current_page",
                        "description": "current_page",
                        "type": "integer"
                    },
                    "from": {
                        "title": "from",
                        "description": "from",
                        "type": "integer"
                    },
                    "last_page": {
                        "title": "last_page",
                        "description": "last_page",
                        "type": "integer"
                    },
                    "path": {
                        "title": "path",
                        "description": "path",
                        "type": "string"
                    },
                    "per_page": {
                        "title": "per_page",
                        "description": "per_page",
                        "type": "integer"
                    },
                    "to": {
                        "title": "to",
                        "description": "to",
                        "type": "integer"
                    },
                    "total": {
                        "title": "total",
                        "description": "total",
                        "type": "integer"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Pagination"
                }
            },
            "GameRequest": {
                "title": "Game request",
                "description": "Game request body data",
                "required": [
                    "name",
                    "studio"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Название",
                        "type": "string",
                        "example": "A nice Game"
                    },
                    "studio": {
                        "title": "studio",
                        "description": "Название",
                        "type": "string",
                        "example": "This is new Game's company"
                    },
                    "genres": {
                        "title": "genres",
                        "description": "Жанры",
                        "type": "array",
                        "items": {
                            "title": "genre.id",
                            "type": "int"
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}